+++ dockstore-webservice/src/main/resources/openapi3/openapi.yaml
@@ -8507,27 +8507,6 @@ paths:
       - workflows
 components:
   schemas:
-    AggregatedExecution:
-      type: object
-      allOf:
-      - $ref: '#/components/schemas/Metrics'
-      - type: object
-        properties:
-          additionalAggregatedMetrics:
-            type: object
-            additionalProperties:
-              type: object
-              description: Additional aggregated metrics
-            description: Additional aggregated metrics
-          executionId:
-            type: string
-            description: User-provided ID of the execution. This ID is used to identify
-              the execution when updating the execution
-      deprecated: true
-      description: Aggregated metrics of multiple executions on a platform
-      required:
-      - executionId
-      - executionStatusCount
     Alias:
       type: object
       properties:
@@ -9608,7 +9587,7 @@ components:
           deprecated: true
           description: List of aggregated executions to submit
           items:
-            $ref: '#/components/schemas/AggregatedExecution'
+            $ref: '#/components/schemas/Metrics'
         runExecutions:
           type: array
           description: List of workflow run executions to submit
@@ -10024,10 +10003,22 @@ components:
       type: object
       description: Aggregated metrics associated with entry versions
       properties:
+        additionalAggregatedMetrics:
+          type: object
+          additionalProperties:
+            type: object
+            deprecated: true
+            description: Additional aggregated metrics
+          deprecated: true
+          description: Additional aggregated metrics
         cost:
           $ref: '#/components/schemas/CostMetric'
         cpu:
           $ref: '#/components/schemas/CpuMetric'
+        executionId:
+          type: string
+          description: User-provided ID of the execution. This ID is used to identify
+            the execution when updating the execution
         executionStatusCount:
           $ref: '#/components/schemas/ExecutionStatusMetric'
         executionTime:
@@ -10041,6 +10032,7 @@ components:
         validationStatus:
           $ref: '#/components/schemas/ValidationStatusMetric'
       required:
+      - executionId
       - executionStatusCount
     Notebook:
       type: object
